+++ noodles-vcf/src/record/genotypes.rs
@@ -14,7 +14,10 @@ use self::genotype::field;
 
 /// VCF record genotypes.
 #[derive(Clone, Debug, Default, PartialEq)]
-pub struct Genotypes(Vec<Genotype>);
+pub struct Genotypes {
+    keys: Keys,
+    genotypes: Vec<Genotype>,
+}
 
 impl Genotypes {
     /// Creates VCF record genotypes.
@@ -22,11 +25,16 @@ impl Genotypes {
     /// # Examples
     ///
     /// ```
-    /// use noodles_vcf::record::Genotypes;
-    /// let genotypes = Genotypes::new(Vec::new());
+    /// use noodles_vcf::record::{genotypes::Keys, Genotypes};
+    /// let genotypes = Genotypes::new(Keys::default(), Vec::new());
     /// ```
-    pub fn new(genotypes: Vec<Genotype>) -> Self {
-        Self(genotypes)
+    pub fn new(keys: Keys, genotypes: Vec<Genotype>) -> Self {
+        Self { keys, genotypes }
+    }
+
+    /// Returns the genotypes keys.
+    pub fn keys(&self) -> &Keys {
+        &self.keys
     }
 
     /// Returns the VCF record genotype value.
@@ -41,19 +49,19 @@ impl Deref for Genotypes {
     type Target = Vec<Genotype>;
 
     fn deref(&self) -> &Self::Target {
-        &self.0
+        &self.genotypes
     }
 }
 
 impl DerefMut for Genotypes {
     fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
+        &mut self.genotypes
     }
 }
 
 impl fmt::Display for Genotypes {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        for (i, genotype) in self.0.iter().enumerate() {
+        for (i, genotype) in self.genotypes.iter().enumerate() {
             if i > 0 {
                 write!(f, "\t")?;
             }
@@ -72,13 +80,13 @@ mod tests {
     #[test]
     fn test_genotypes() -> Result<(), Box<dyn std::error::Error>> {
         let keys = "GT:GQ".parse()?;
-
-        let genotypes = Genotypes::new(vec![
+        let genotypes = vec![
             Genotype::from_str_format("0|0:7", &keys)?,
             Genotype::from_str_format("./.:20", &keys)?,
             Genotype::from_str_format("1/1:1", &keys)?,
             Genotype::from_str_format(".", &keys)?,
-        ]);
+        ];
+        let genotypes = Genotypes::new(keys, genotypes);
 
         let actual = genotypes.genotypes();
         let expected = Ok(vec![
@@ -100,10 +108,13 @@ mod tests {
             Field,
         };
 
-        let genotypes = Genotypes::new(vec![Genotype::try_from(vec![
-            Field::new(Key::Genotype, Some(Value::String(String::from("0|0")))),
-            Field::new(Key::ConditionalGenotypeQuality, Some(Value::Integer(13))),
-        ])?]);
+        let genotypes = Genotypes::new(
+            Keys::try_from(vec![Key::Genotype, Key::ConditionalGenotypeQuality])?,
+            vec![Genotype::try_from(vec![
+                Field::new(Key::Genotype, Some(Value::String(String::from("0|0")))),
+                Field::new(Key::ConditionalGenotypeQuality, Some(Value::Integer(13))),
+            ])?],
+        );
 
         assert_eq!(genotypes.to_string(), "0|0:13");