+++ noodles-vcf/src/record/parser.rs
@@ -78,23 +78,22 @@ pub fn parse(s: &str, header: &Header) -> Result<Record, ParseError> {
     let info = parse_string(&mut fields, Field::Info)
         .and_then(|s| Info::try_from_str(s, header.infos()).map_err(ParseError::InvalidInfo))?;
 
-    let format = match fields.next() {
+    let mut format = match fields.next() {
         Some(s) => genotypes::Keys::try_from_str(s, header.formats())
             .map(Some)
             .map_err(ParseError::InvalidFormat)?,
         None => None,
     };
 
-    let genotypes = format
-        .as_ref()
-        .map(|f| {
-            fields
-                .map(|s| genotypes::Genotype::from_str_format(s, f))
-                .collect::<Result<_, _>>()
-                .map_err(ParseError::InvalidGenotype)
-        })
-        .unwrap_or_else(|| Ok(Vec::new()))
-        .map(Genotypes::new)?;
+    let genotypes = if let Some(keys) = format.take() {
+        fields
+            .map(|s| genotypes::Genotype::from_str_format(s, &keys))
+            .collect::<Result<_, _>>()
+            .map(|genotypes| Genotypes::new(keys, genotypes))
+            .map_err(ParseError::InvalidGenotype)?
+    } else {
+        Genotypes::default()
+    };
 
     Ok(Record {
         chromosome: chrom,
@@ -105,7 +104,6 @@ pub fn parse(s: &str, header: &Header) -> Result<Record, ParseError> {
         quality_score: qual,
         filters: filter,
         info,
-        format,
         genotypes,
     })
 }
@@ -168,7 +166,6 @@ mod tests {
         assert_eq!(record.quality_score().map(f32::from), Some(5.8));
         assert_eq!(record.filters(), Some(&Filters::Pass));
         assert_eq!(record.info().len(), 1);
-        assert!(record.format().is_none());
         assert!(record.genotypes().is_empty());
 
         Ok(())
@@ -176,35 +173,27 @@ mod tests {
 
     #[test]
     fn test_from_str_with_genotype_info() -> Result<(), Box<dyn std::error::Error>> {
-        use genotypes::genotype;
+        use genotypes::{
+            genotype::{
+                field::{Key, Value},
+                Field,
+            },
+            Genotype,
+        };
 
         let s = "chr1\t13\tnd0\tATCG\tA\t5.8\tPASS\tSVTYPE=DEL\tGT:GQ\t0|1:13";
         let record: Record = s.parse()?;
 
-        let expected = genotypes::Keys::try_from(vec![
-            genotype::field::Key::Genotype,
-            genotype::field::Key::ConditionalGenotypeQuality,
-        ])?;
+        let keys = genotypes::Keys::try_from(vec![Key::Genotype, Key::ConditionalGenotypeQuality])?;
+        let genotypes = vec![Genotype::try_from(vec![
+            Field::new(Key::Genotype, Some(Value::String(String::from("0|1")))),
+            Field::new(Key::ConditionalGenotypeQuality, Some(Value::Integer(13))),
+        ])?];
 
-        assert_eq!(record.format(), Some(&expected));
+        let actual = record.genotypes();
+        let expected = Genotypes::new(keys, genotypes);
 
-        let genotypes = record.genotypes();
-
-        assert_eq!(genotypes.len(), 1);
-
-        let actual: Vec<_> = genotypes[0].values().cloned().collect();
-        let expected = vec![
-            genotype::Field::new(
-                genotype::field::Key::Genotype,
-                Some(genotype::field::Value::String(String::from("0|1"))),
-            ),
-            genotype::Field::new(
-                genotype::field::Key::ConditionalGenotypeQuality,
-                Some(genotype::field::Value::Integer(13)),
-            ),
-        ];
-
-        assert_eq!(actual, expected);
+        assert_eq!(actual, &expected);
 
         Ok(())
     }